app/requirements.txt
alembic==1.9.1 ; python_version >= "3.10" and python_version < "4.0"
autoflake==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
black==22.12.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0"
cssbeautifier==1.14.7 ; python_version >= "3.10" and python_version < "4.0"
djlint==1.19.9 ; python_version >= "3.10" and python_version < "4.0"
dominate==2.7.0 ; python_version >= "3.10" and python_version < "4.0"
editorconfig==0.12.3 ; python_version >= "3.10" and python_version < "4.0"
flask-bootstrap==3.3.7.1 ; python_version >= "3.10" and python_version < "4.0"
flask-login==0.6.2 ; python_version >= "3.10" and python_version < "4.0"
flask-migrate==4.0.0 ; python_version >= "3.10" and python_version < "4.0"
flask-sqlalchemy==3.0.2 ; python_version >= "3.10" and python_version < "4.0"
flask==2.2.2 ; python_version >= "3.10" and python_version < "4.0"
greenlet==2.0.1 ; python_version >= "3.10" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "4.0"
html-tag-names==0.1.2 ; python_version >= "3.10" and python_version < "4.0"
html-void-elements==0.1.0 ; python_version >= "3.10" and python_version < "4.0"
importlib-metadata==5.2.0 ; python_version >= "3.10" and python_version < "4.0"
install==1.3.5 ; python_version >= "3.10" and python_version < "4.0"
isort==5.11.4 ; python_version >= "3.10" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
jsbeautifier==1.14.7 ; python_version >= "3.10" and python_version < "4.0"
mako==1.2.4 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.1 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==0.4.3 ; python_version >= "3.10" and python_version < "4.0"
pathspec==0.10.3 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==2.6.0 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.5 ; python_version >= "3.10" and python_version < "4.0"
pyflakes==3.0.1 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
regex==2022.10.31 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==1.4.45 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.10" and python_full_version < "3.11.0a7"
tqdm==4.64.1 ; python_version >= "3.10" and python_version < "4.0"
visitor==0.1.3 ; python_version >= "3.10" and python_version < "4.0"
werkzeug==2.2.2 ; python_version >= "3.10" and python_version < "4.0"
zipp==3.11.0 ; python_version >= "3.10" and python_version < "4.0"
